+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Added
+- Added the `php craft tinymce/config/regenerate-default` console command
+- Added the `php craft tinymce/config/regenerate-full` console command
+- Added the `php craft tinymce/config/regenerate-simple` console command
+- Added `spicyweb\tinymce\console\controllers\ConfigController`
+- Added `spicyweb\tinymce\services\ConfigService`
+
+### Changed
+- When installed, the plugin now saves `Default.json`, `Full.json` and `Simple.json` TinyMCE config files in a Craft install's `config/tinymce` directory, if the files didn't already exist
+
 ## 1.0.1 - 2022-09-26
 
 ### Fixed